requirements.txt
python-telegram-bot>=21.0.1
sentence-transformers>=2.6.1
groq>=0.5.0
python-dotenv>=1.0.1
fastapi>=0.115.0
uvicorn>=0.30.0
qdrant-client>=1.11.0
pydantic-settings>=2.4.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
